requirements.txt
numpy==1.24.4
pandas==1.5.3
Pillow==11.0.0
scipy==1.14.1
timm==1.0.11
torch==2.2.1+cu121
torchvision==0.17.1+cu121
tqdm==4.66.1